Pulled up code into a superclass that will be useful for other UI main classes.
